requirements.txt
django
mysqlclient
djangorestframework